refactor: Separate queries related to Feed feature
