+++ src/features/core/api/loaders.ts
@@ -1,19 +1,11 @@
 import { QueryClient } from "@tanstack/react-query";
 import { currentUserQuery } from "../utils/auth";
 import { allPostsQuery } from "../../explore/api/queries";
+import { ActionFunctionArgs, ParamParseKey, Params } from "react-router-dom";
 import {
-  ActionFunctionArgs,
-  LoaderFunctionArgs,
-  ParamParseKey,
-  Params,
-} from "react-router-dom";
-import {
-  followingPostsQuery,
   followQuery,
-  forYouPostsQuery,
   likeQuery,
   postQuery,
-  recommendedUsersQuery,
   userPostsQuery,
   usersByUsernameQuery,
 } from "./queries";
@@ -48,36 +40,6 @@ export const appLoader = (queryClient: QueryClient) => async () => {
       return null;
     });
 };
-
-export const feedLoader =
-  (queryClient: QueryClient) =>
-  async ({ request }: LoaderFunctionArgs) => {
-    const variant = new URL(request.url).searchParams.get("variant");
-    const postsQuery =
-      variant === "following" ? followingPostsQuery : forYouPostsQuery;
-
-    const postsPromise = queryClient
-      .ensureQueryData(postsQuery())
-      .then((data) => {
-        return data;
-      })
-      .catch(() => {
-        return [];
-      });
-
-    const recUsersPromise = queryClient
-      .ensureQueryData(recommendedUsersQuery())
-      .then((data) => {
-        return data;
-      })
-      .catch(() => {
-        return [];
-      });
-
-    return Promise.all([recUsersPromise, postsPromise]).then((values) => {
-      return { posts: values[0], userRecommendations: values[1] };
-    });
-  };
 
 export const exploreLoader = (queryClient: QueryClient) => async () => {
   const query = allPostsQuery();
