+++ src/features/core/api/queries.ts
@@ -4,7 +4,6 @@ import ObjectId from "../types/objectId";
 import Like from "../types/like";
 import Post from "../types/post";
 import User from "../types/user";
-import UserRecommendation from "../types/userRecommendation";
 
 const getFollow = async (targetId: ObjectId) => {
   return api.get<Follow>(`/follows/${targetId}`).then((response) => {
@@ -64,18 +63,6 @@ const getPost = async (id: ObjectId) => {
   });
 };
 
-const getForYouPosts = async () => {
-  return api.get<Post[]>(`/posts/for-you`).then((response) => {
-    return response.data;
-  });
-};
-
-const getFollowingPosts = async () => {
-  return api.get<Post[]>(`/posts/following`).then((response) => {
-    return response.data;
-  });
-};
-
 const getUserPosts = async (userId: ObjectId) => {
   return api.get<Post[]>(`/posts?user=${userId}`).then((response) => {
     return response.data;
@@ -97,27 +84,9 @@ const getUsersByUsername = async (username: string, partial: boolean) => {
   });
 };
 
-const getReccomendedUsers = async () => {
-  return api
-    .get<UserRecommendation[]>("/users/recommended")
-    .then((response) => {
-      return response.data;
-    });
-};
-
 const postQuery = (postId: ObjectId) => ({
   queryKey: ["posts", postId],
   queryFn: async () => getPost(postId),
-});
-
-const forYouPostsQuery = () => ({
-  queryKey: ["posts", "forYou"],
-  queryFn: async () => getForYouPosts(),
-});
-
-const followingPostsQuery = () => ({
-  queryKey: ["posts", "following"],
-  queryFn: async () => getFollowingPosts(),
 });
 
 const likeQuery = (postId: ObjectId) => ({
@@ -140,13 +109,6 @@ const userPostsQuery = (userId: ObjectId) => ({
   queryFn: async () => getUserPosts(userId),
 });
 
-const recommendedUsersQuery = () => ({
-  queryKey: ["users", "recommended"],
-  queryFn: async () => getReccomendedUsers(),
-  retryDelay: (attempt: number) => attempt * 1000,
-  staleTime: 1000 * 60 * 5,
-});
-
 export {
   getFollow,
   getFollowers,
@@ -157,12 +119,9 @@ export {
   postLike,
   deleteLike,
   postQuery,
-  forYouPostsQuery,
-  followingPostsQuery,
   deletePost,
   likeQuery,
   followQuery,
   usersByUsernameQuery,
   userPostsQuery,
-  recommendedUsersQuery,
 };
