+++ src/features/feed/hooks/useFollowingPostsQuery.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { followingPostsQuery } from "../../core/api/queries";
+import { followingPostsQuery } from "../api/queries";
 
 export default function useFollowingPostsQuery() {
   return useQuery(followingPostsQuery());
