+++ src/features/feed/hooks/useRecommendedUsersQuery.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { recommendedUsersQuery } from "../../core/api/queries";
+import { recommendedUsersQuery } from "../api/queries";
 
 export default function useRecommendedUsersQuery() {
   return useQuery(recommendedUsersQuery());
